refactor pipline

+++ ops/common/alpha/results/compliance.py
@@ -10,13 +10,14 @@ class CompStatus(Enum):
 
 class CompResult:
     def __init__(self, avg_long_pct: np.float64, avg_short_pct: np.float64,
-                 long_count: np.float64, short_count: np.float64,
+                 long_count: int, short_count: int,
                  total_checked: int = 0):
         self.avg_long_pct = avg_long_pct
         self.avg_short_pct = avg_short_pct
         self.long_count = long_count
         self.short_count = short_count
         self.total_checked = total_checked
+
 
 class CompResults:
     def __init__(self):
